refactor(gear-programs): store eth-events* program as a template and use it where appropriate
