+++ gear-programs/eth-events-deneb/build.rs
@@ -19,5 +19,8 @@ fn main() {
 
     let mut idl_path = PathBuf::from(bin_path);
     idl_path.set_extension("idl");
-    sails_idl_gen::generate_idl_to_file::<eth_events_deneb_app::Program>(idl_path).unwrap();
+    sails_idl_gen::generate_idl_to_file::<eth_events_deneb_app::EthereumEventClientProgram>(
+        idl_path,
+    )
+    .unwrap();
 }
